FAC-135 fix: break TieredPipelineSchedulerJob ↔ orchestrator import cycle
The Phase B wiring created a circular import at module-load time:

  pipeline-orchestrator.service.ts
    → lib/next-scheduled-run.ts
      → crons/jobs/analysis-jobs/tiered-pipeline-scheduler.job.ts
        → pipeline-orchestrator.service.ts

When Nest resolved TieredPipelineSchedulerJob's constructor, the
PipelineOrchestratorService class reference was still undefined,
manifesting as:

  UndefinedDependencyException: Nest can't resolve dependencies of
  the TieredPipelineSchedulerJob (?, SqlEntityManager, SchedulerRegistry).

Extract TIERED_SCHEDULER_CRON_NAMES + TIERED_SCHEDULER_CRON_EXPRS into
their own constants module (tiered-scheduler.constants.ts) so
next-scheduled-run.ts pulls only the constants — no value-level import
of the job class. The job re-exports the constants for any back-compat
caller.
